% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ab}
\alias{ab}
\title{Load calulation results from the original AutoBeale program.}
\format{An object of class \code{data.frame} with 5 rows and 11 columns.}
\usage{
ab
}
\description{
A dataset that contains load calculation results as generated by the original AutoBeale
program for 1997 for the Rocky River, Ohio.
}
\keyword{datasets}
