% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beale_function.R
\name{beale}
\alias{beale}
\title{Calculate pollutant loads by means of the Beale Ratio Estimator.}
\usage{
beale(discharge_cms, concentration_mg_L)
}
\arguments{
\item{discharge_cms}{vector of daily discharge values given in cubic meters per second.}

\item{concentration_mg_L}{paired vector of daily concentration values given in mg_L}
}
\value{
A list with the following items:
\tabular{lll}{
Name \tab Description \cr
load_total_corrected \tab bias-corrected load for entire period of calculation \cr
load_daily_corrected \tab bias-corrected daily load \cr
load_mean_daily \tab mean of daily loads over period of calculation \cr
load_daily \tab vector of individual daily loads \cr
confidence_interval \tab 95\% confidence interval about the total corrected load \cr
discharge_mean \tab mean discharge during calculation period \cr
discharge_mean_sample_days_only \tab mean discharge, including only days for which concentration data exist \cr
mse \tab mean squared error \cr
rmse \tab root mean square of errors \cr
df \tab number of degrees of freedom \cr
n_conc \tab number of concentration data points used in calculation \cr
n_discharge \tab number of discharge values used in calculation \cr
}
}
\description{
This function is an R implementation of the code originally written
in Fortran. The Beale Ratio Estimator has been shown to provide an unbiased load estimate.
}
\details{
Peter Richards, one of the authors of the Fortran implementation, describes it this way:

"The Beale Ratio Estimator is described in Tin (1965). It has been widely used for computation of
substance loadings to receiving waters, particularly in the Great Lakes system. The initial code was
developed by Ken Baun of Wisconsin DNR (1982). It was subsequently used and developed by Kevin McGunagle
at the International Joint Commission (IJC), from whom the current version was received. It was then
developed in the 1990’s by Pete Richards at the National Center for Water Quality Research
at Heidelberg University (www.heidelberg.edu/academiclife/distinctive/ncwqr). Peter developed the code
to work with the Macintosh computer of the time (“Motorola" architecture, which was replaced by
Intel architecture ca. 2005), and included a Graphic User Interface (GUI) and an algorithm that objectively
and sequentially searched out the best stratification given the data, with “best” being defined as
minimizing the root-mean-square error (RMSE) given the data at hand. This is the current and final version.
A Windows version of the code was also developed, with a minimal GUI."
}
\examples{
library(dplyr)
discharge_data <- read.table("ROCKFLOW.DAT", header=TRUE )
concentration_data <- read.table("ROCKNO3.DAT", header=TRUE )
flow_conc_df <- left_join(discharge_data, concentration_data, by=c("date" = "date") )
head( flow_conc_df )
beale( discharge_cms=cfs_to_cms( flow_conc_df$discharge ), concentration_mg_L=flow_conc_df$concentration )
}
