% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{daily_loads}
\alias{daily_loads}
\title{Daily load calculation results from the original AutoBeale program.}
\format{An object of class \code{data.frame} with 52 rows and 7 columns.}
\usage{
daily_loads
}
\description{
A dataset that contains \emph{daily} load calculation results as generated by the original AutoBeale
program for 1997 for the Rocky River, Ohio.
}
\keyword{datasets}
