% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partitions.R
\docType{data}
\name{partitions}
\alias{partitions}
\title{Data structure to hold partition information.}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
partitions
}
\description{
In addition to storing information on the partitions, a number of
functions are associated with the data structure.
}
\section{Slots}{

\describe{
\item{\code{initialize}}{reserve space for the appropriate number of partitions}

\item{\code{get_date_fractions}}{return the date fractions associated with the current set of partitions}

\item{\code{get_date_values}}{return the dates associated with the current set of partitions}
}}

\keyword{datasets}
