% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{qcdata}
\alias{qcdata}
\title{Discharge and nitrate-nitrite concentration data for Rocky River, Ohio.}
\format{A data frame with 365 rows and 3 variables:
\describe{
\item{date}{date, as an R data class}
\item{discharge}{daily mean discharge, in cubic feet per second}
\item{concentration}{concentration of nitrate-nitrite, in milligrams per liter}
}}
\usage{
qcdata
}
\description{
A dataset that contains discharge and nitrate-nitrite concentration data
from 1997 for the Rocky River, Ohio.
}
\keyword{datasets}
