% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratum.R
\docType{data}
\name{stratum}
\alias{stratum}
\title{Data structure to hold information for a single stratum.}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
stratum
}
\description{
Data structure to holds load and concentration data for
an individual stratum, and also makes the call to \code{beale} to estimate
the stratum's load.
}
\section{Slots}{

\describe{
\item{\code{update}}{adjusts the bounds of the current stratum based on the input
start and end date; subsets appropriate discharge and concentration data}

\item{\code{calc_load}}{calculate the load for the stratum defined by this data structure}
}}

\keyword{datasets}
